+++ src/components/Insights/InsightsCatalog/PromotionCard/EarlyAccessPromotionCard/EarlyAccessPromotionCard.stories.tsx
@@ -1,10 +1,10 @@
 import { Meta, StoryObj } from "@storybook/react";
-import { UdemyCoursePromotionCard } from ".";
+import { EarlyAccessPromotionCard } from ".";
 
 // More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction
-const meta: Meta<typeof UdemyCoursePromotionCard> = {
+const meta: Meta<typeof EarlyAccessPromotionCard> = {
   title: "Insights/InsightsCatalog/PromotionCard/EarlyAccessPromotionCard",
-  component: UdemyCoursePromotionCard,
+  component: EarlyAccessPromotionCard,
   parameters: {
     // More on how to position stories at: https://storybook.js.org/docs/react/configure/story-layout
     layout: "fullscreen"
