+++ src/components/Insights/InsightsCatalog/PromotionSection/types.ts
@@ -1,0 +1,1 @@
+export type PromotionType = "udemy" | "early-access";
